refactor: migrate playingfield package into game domain and application
