+++ src/main/java/com/codenames/codenames/backend/game/domain/Card.java
@@ -1,4 +1,4 @@
-package com.codenames.codenames.backend.playingfield;
+package com.codenames.codenames.backend.game.domain;
 
 import com.codenames.codenames.backend.utility.Color;
 import lombok.Getter;
