+++ src/main/java/com/codenames/codenames/backend/lobby/api/LobbySocketController.java
@@ -2,8 +2,8 @@ package com.codenames.codenames.backend.lobby.api;
 
 import com.codenames.codenames.backend.lobby.api.dto.JoinMessage;
 import com.codenames.codenames.backend.lobby.domain.Player;
+import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.lobby.services.LobbyService;
-import com.codenames.codenames.backend.playingfield.GameService;
 import com.codenames.codenames.backend.shared.websocket.SessionRegistry;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
