+++ src/main/java/com/codenames/codenames/backend/lobby/services/LobbyService.java
@@ -1,10 +1,10 @@
 package com.codenames.codenames.backend.lobby.services;
 
 import com.codenames.codenames.backend.chat.ChatService;
+import com.codenames.codenames.backend.game.application.GameService;
 import com.codenames.codenames.backend.lobby.Lobby;
 import com.codenames.codenames.backend.lobby.domain.Player;
 import com.codenames.codenames.backend.lobby.dto.PlayerDto;
-import com.codenames.codenames.backend.playingfield.GameService;
 import com.codenames.codenames.backend.utility.Role;
 import com.codenames.codenames.backend.utility.Team;
 import java.util.List;
