+++ src/test/java/com/codenames/codenames/backend/recovery/SystemStateRecoveryServiceTest.java
@@ -7,15 +7,15 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.codenames.codenames.backend.chat.ChatService;
 import com.codenames.codenames.backend.clue.ClueValidationService;
+import com.codenames.codenames.backend.game.application.CardGenerator;
+import com.codenames.codenames.backend.game.application.GameService;
+import com.codenames.codenames.backend.game.domain.GameManager;
+import com.codenames.codenames.backend.game.domain.GameManagerFactory;
 import com.codenames.codenames.backend.game.dto.ClueDto;
 import com.codenames.codenames.backend.lobby.Lobby;
 import com.codenames.codenames.backend.lobby.dto.PlayerDto;
 import com.codenames.codenames.backend.lobby.services.LobbyCodeGenerator;
 import com.codenames.codenames.backend.lobby.services.LobbyService;
-import com.codenames.codenames.backend.playingfield.CardGenerator;
-import com.codenames.codenames.backend.playingfield.GameManager;
-import com.codenames.codenames.backend.playingfield.GameManagerFactory;
-import com.codenames.codenames.backend.playingfield.GameService;
 import com.codenames.codenames.backend.recovery.snapshot.SystemSnapshot;
 import com.codenames.codenames.backend.serialization.CardDataTransferObject;
 import com.codenames.codenames.backend.serialization.DataTransferObjectService;
