+++ src/test/java/com/codenames/codenames/backend/serialization/DataTransferObjectServiceTest.java
@@ -6,8 +6,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.codenames.codenames.backend.clue.Clue;
-import com.codenames.codenames.backend.playingfield.Card;
-import com.codenames.codenames.backend.playingfield.GameManager;
+import com.codenames.codenames.backend.game.domain.Card;
+import com.codenames.codenames.backend.game.domain.GameManager;
 import com.codenames.codenames.backend.utility.Color;
 import com.codenames.codenames.backend.utility.Role;
 import com.codenames.codenames.backend.utility.Team;
