+++ src/test/java/com/codenames/codenames/backend/serialization/SerializationJsonTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.codenames.codenames.backend.game.dto.ClueDto;
-import com.codenames.codenames.backend.playingfield.Card;
+import com.codenames.codenames.backend.game.domain.Card;
 import com.codenames.codenames.backend.utility.Color;
 import com.codenames.codenames.backend.utility.Role;
 import com.codenames.codenames.backend.utility.Team;
